+++ kotest-assertions/kotest-assertions-json/src/commonMain/kotlin/io/kotest/assertions/json/JsonMatchers.kt
@@ -60,24 +60,20 @@ fun matchJson(expected: String?) = object : Matcher<String?> {
  * regardless of order.
  *
  */
-fun String.shouldEqualJson(expected: String, mode: CompareMode, order: CompareOrder) {
+fun String.shouldEqualJson(expected: String, mode: CompareMode, order: CompareOrder) =
+   this.shouldEqualJson(expected, legacyOptions(mode, order))
+
+fun String.shouldEqualJson(expected: String, options: CompareJsonOptions) {
    val (e, a) = parse(expected, this)
-   a should equalJson(e, mode, order, FieldComparison.Exact)
+   a should equalJson(e, options)
 }
 
-fun String.shouldNotEqualJson(expected: String, mode: CompareMode, order: CompareOrder) {
+fun String.shouldNotEqualJson(expected: String, mode: CompareMode, order: CompareOrder) =
+   this.shouldNotEqualJson(expected, legacyOptions(mode, order))
+
+fun String.shouldNotEqualJson(expected: String, options: CompareJsonOptions) {
    val (e, a) = parse(expected, this)
-   a shouldNot equalJson(e, mode, order, FieldComparison.Exact)
-}
-
-fun String.shouldEqualSpecifiedJson(expected: String, mode: CompareMode, order: CompareOrder) {
-   val (e,a) = parse(expected, this)
-   a should equalJson(e, mode, order, FieldComparison.IgnoreExtra)
-}
-
-fun String.shouldNotEqualSpecifiedJson(expected: String, mode: CompareMode, order: CompareOrder) {
-   val (e,a) = parse(expected, this)
-   a shouldNot equalJson(e, mode, order, FieldComparison.IgnoreExtra)
+   a shouldNot equalJson(e, options)
 }
 
 internal fun parse(expected: String, actual: String): Pair<JsonTree, JsonTree> {